+++ src/helpers/index.js
@@ -1,0 +1,5 @@
+const isExist = require('./is-exist');
+
+module.exports = {
+  isExist,
+};