+++ src/helpers/is-exist.js
@@ -1,0 +1,5 @@
+module.exports = isExist;
+
+function isExist(value) {
+  return (value || value === 0);
+}